+++ generic_k8s_webhook/config_parser/common.py
@@ -1,0 +1,2 @@
+class ParsingException(Exception):
+    pass
